Mv `StreamReadable` into `crate::utils`
